Hotfix/sr2

+++ src/main/java/com/Heart2Hub/Heart2Hub_Backend/service/FacilityBookingService.java
@@ -4,8 +4,10 @@ import com.Heart2Hub.Heart2Hub_Backend.entity.Facility;
 import com.Heart2Hub.Heart2Hub_Backend.entity.FacilityBooking;
 import com.Heart2Hub.Heart2Hub_Backend.entity.Staff;
 import com.Heart2Hub.Heart2Hub_Backend.exception.OverlappingBookingException;
+import com.Heart2Hub.Heart2Hub_Backend.exception.UnableToDeleteFacilityException;
 import com.Heart2Hub.Heart2Hub_Backend.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cglib.core.Local;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,17 +40,11 @@ public class FacilityBookingService {
   }
 
   public void updateBooking(Long facilityBookingId, Long facilityId) {
-    System.out.println("here1");
     FacilityBooking facilityBooking = facilityBookingRepository.findById(facilityBookingId).get();
-    System.out.println("here2");
     Facility oldFacility = facilityBooking.getFacility();
-    System.out.println("here3");
     oldFacility.getListOfFacilityBookings().remove(facilityBooking);
-    System.out.println("here4");
     Facility facility = facilityRepository.findById(facilityId).get();
-    System.out.println("here5");
     facility.getListOfFacilityBookings().add(facilityBooking);
-    System.out.println("here6");
     facilityBooking.setFacility(facility);
   }
 
@@ -124,6 +120,10 @@ public class FacilityBookingService {
       throw new OverlappingBookingException("Start Time same as End Time!");
     }
 
+    if (existingBooking.getShift() != null) {
+      throw new UnableToDeleteFacilityException("Cannot change booking associated with a shift");
+    }
+
     existingBooking.setComments(comments);
     existingBooking.setStartDateTime(startDateTime);
     existingBooking.setEndDateTime(endDateTime);
@@ -136,9 +136,18 @@ public class FacilityBookingService {
     FacilityBooking facilityBooking = getFacilityBookingById(id);
 
     Staff s = staffRepository.findByUsername(facilityBooking.getStaffUsername()).get();
+if (facilityBooking.getShift() != null) {
+  throw new UnableToDeleteFacilityException("Cannot delete booking associated with a shift");
+}
     s.getListOfFacilityBookings().remove(facilityBooking);
     staffRepository.save(s);
 
     facilityBookingRepository.delete(facilityBooking);
+  }
+
+  public List<FacilityBooking> getAllFacilityBookingsWithinTime(String name, LocalDateTime start, LocalDateTime end) {
+    start = start.plusMinutes(1L);
+    end = end.minusMinutes(1L);
+    return facilityBookingRepository.findAllByFacilityNameAndStartDateTimeBetween(name, start, end);
   }
 }
